Check for valid id and return not found if not valid/existing.
Comply with standard REST response protocol.

+++ src/workflow_collection.py
@@ -1,8 +1,16 @@
 from pymongo import MongoClient
 from bson.objectid import ObjectId
-
+from bson.errors import InvalidId
 
 access_url = 'workflow-editor-datastore:27017'
+
+def is_valid_object_id(obj_id):
+    try:
+        wid = ObjectId(obj_id)
+        return True
+    except InvalidId:
+        return False
+
 
 
 class WorkflowCollection:
@@ -13,9 +21,11 @@ class WorkflowCollection:
         self.workflow_collection = self.db.workflow_collection
 
     def get_one_workflow(self, wid):
+        if not is_valid_object_id(wid):
+            raise ValueError
         workflow = self.workflow_collection.find_one(ObjectId(wid))
         if not workflow:
-            return 'No results'
+            raise ValueError
         workflow['_id'] = str(workflow['_id'])
         return workflow
 
@@ -32,13 +42,21 @@ class WorkflowCollection:
         return str(wid)
 
     def update_one_workflow(self, wid, updates):
-        updates = {'$set': updates}
+        if not is_valid_object_id(wid):
+            raise ValueError
 
-        self.workflow_collection.update_one({'_id': ObjectId(wid)}, updates)
+        updates = {'$set': updates}
+        update_res = self.workflow_collection.update_one({'_id': ObjectId(wid)}, updates)
+        if update_res.matched_count == 0:
+            raise ValueError
 
     def delete_one_workflow(self, wid):
-        self.workflow_collection.delete_one({'_id': ObjectId(wid)})
-        return True
+        if not is_valid_object_id(wid):
+            raise ValueError
+        del_res = self.workflow_collection.delete_one({'_id': ObjectId(wid)})
+        if del_res.deleted_count == 0:
+            raise ValueError
+
 
     def delete_all_workflows(self):
         self.workflow_collection.delete_many({})
